Disaster/energy-orchestrator-platform/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
aiohttp==3.9.1

# AI/ML
langchain==0.0.350
openai==1.3.7
anthropic==0.7.7
torch==2.1.1
torch-geometric==2.4.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2

# IoT & Streaming
paho-mqtt==1.6.1
kafka-python==2.0.2
aiokafka==0.10.0

# Ontology
rdflib==7.0.0
SPARQLWrapper==2.0.0

# Time Series
# timescaledb==1.0.1  # TimescaleDB is a PostgreSQL extension, not a Python package

# Weather API
# openweathermap-api==0.1.0  # Package not available, using requests instead
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
geopy==2.4.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2



